# openapi.yaml
openapi: 3.0.3
info:
  title: POD Designer Demo API
  version: 0.1.0
servers:
  - url: http://localhost:4000
    description: Backend API
  - url: http://localhost:8000
    description: Measurement Service
paths:
...
  /api/uploads:
    post:
      summary: Upload an image to S3/MinIO
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                metadata:
                  type: string
                  description: JSON string (e.g. exif)
      responses:
        '200':
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
  /api/uploads/process:
    post:
      summary: Enqueue image-processing job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                key:
                  type: string
                actions:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Job enqueued
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: integer
  /api/uploads/jobs/{id}:
    get:
      summary: Query job status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Job info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
...
components:
  schemas:
    UploadResponse:
      type: object
      properties:
        url:
          type: string
        key:
          type: string
...
    Job:
      type: object
      properties:
        id: { type: integer }
        type: { type: string }
        payload: { type: object }
        status: { type: string, enum: [queued, processing, done, failed] }
        result:
          type: object
          properties:
            parts:
              type: array
              items:
                $ref: '#/components/schemas/Part'
            overallConfidence:
              type: number
    Part:
      type: object
      properties:
        name: { type: string }
        polygon:
          type: array
          items:
            type: object
            properties:
              x: { type: number }
              y: { type: number }
        rectifiedKey: { type: string }
        rectifiedUrl: { type: string }
        maskKey: { type: string }
        maskUrl: { type: string }
        rectifiedPx:
          type: object
          properties:
            w: { type: number }
            h: { type: number }
        physicalSize:
          type: object
          properties:
            width_m: { type: number }
            height_m: { type: number }
            area_m2: { type: number }
...